requirements.txt
matplotlib>=2.2.3
numpy>=1.14.3
shapely>=1.6.4.post1
descartes>=1.1.0
pybind11>=2.2.4
fitsio>=1.0.5
sdss-coordio>=1.1.0
psutil>=5.8.0
